+++ HISTORY.rst
@@ -87,4 +87,4 @@ robosapiensIO Preview
 ~~~~~~~~~~~~~~~~~~
 
 * Update swc2main
-* swc2code updated (user todo added))
+* swc2code updated (user todo added)